Refactor: Restructure project and add new features
This commit restructures the project by removing unused files and classes.
It also introduces new fragments (StatusFragment and ProfileFragment).
The view pager adapter now uses the StatusFragment as the default.
The main activity layout has been updated to use new fragments.
The MainFragment was renamed to StatusFragment, and a new ProfileFragment was created.
Added `changePassword` and `sendResetPasswordEmail` in API Client and `ChangePasswordRequest` and `ResetPasswordRequest` in model.
The project is also updated to disable night mode.
Login also includes reset password functionality.